added decimal fields

+++ final_fct_nfct_deal/models.py
@@ -18,9 +18,9 @@ class FinalFctNfctDeal(models.Model):
     agency_name_ref     = models.ForeignKey(AgencyDetail,on_delete = models.CASCADE,default = 'default')
     agency_contact_ref  = models.ForeignKey(AgencyContact,on_delete = models.CASCADE,default = 'default')
     brand_name_ref      = models.ForeignKey(CustomerName,on_delete = models.CASCADE, default = 'default', related_name = 'brand')
-    fct_total           = models.IntegerField(null=True,blank=True,default='0')
-    nfct_total          = models.IntegerField(null=True,blank=True,default='0')
-    grandtotal          = models.IntegerField(null=True,blank=True)
+    fct_total           = models.DecimalField(null=True,blank=True,default='0', max_digits=12, decimal_places=2)
+    nfct_total          = models.DecimalField(null=True,blank=True,default='0', max_digits=12, decimal_places=2)
+    grandtotal          = models.DecimalField(null=True,blank=True, max_digits=12, decimal_places=2)
 
     def __str__(self):
         return self.deal_id
